backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0
anthropic==0.34.2
google-api-python-client==2.147.0
python-dotenv==1.0.1
sse-starlette==2.1.3
lxml==5.3.0
python-docx==1.1.2
pytrends==4.9.2
googlesearch-python==1.2.5
